Add feature flag for cors traits
