+++ htsget-test-utils/src/lib.rs
@@ -1,5 +1,7 @@
-#[cfg(feature = "test-traits")]
-pub mod test_traits;
+#[cfg(feature = "http-tests")]
+pub mod http_tests;
+#[cfg(feature = "cors-tests")]
+pub mod cors_tests;
 #[cfg(feature = "server-tests")]
 pub mod server_tests;
 pub mod util;
