--- htsget-test-utils/src/test_traits.rs
@@ -1,48 +1,0 @@
-use http::HeaderMap;
-use serde::de;
-
-/// Represents a http header.
-#[derive(Debug)]
-pub struct Header<T: Into<String>> {
-  pub name: T,
-  pub value: T,
-}
-
-impl<T: Into<String>> Header<T> {
-  pub fn into_tuple(self) -> (String, String) {
-    (self.name.into(), self.value.into())
-  }
-}
-
-/// Represents a http response.
-#[derive(Debug)]
-pub struct Response {
-  pub status: u16,
-  pub headers: HeaderMap,
-  pub body: Vec<u8>,
-  pub expected_url_path: String,
-}
-
-impl Response {
-  pub fn new(status: u16, headers: HeaderMap, body: Vec<u8>, expected_url_path: String) -> Self {
-    Self {
-      status,
-      headers,
-      body,
-      expected_url_path,
-    }
-  }
-
-  /// Deserialize the body from a slice.
-  pub fn deserialize_body<T>(&self) -> Result<T, serde_json::Error>
-    where
-      T: de::DeserializeOwned,
-  {
-    serde_json::from_slice(&self.body)
-  }
-
-  /// Check if status code is success.
-  pub fn is_success(&self) -> bool {
-    300 > self.status && self.status >= 200
-  }
-}
